Wesnoth mod: display last post subject in forum lists
This is a custom modification that's strongly derivative of the
functionality suggested for phpBB 3.1, although it's presumably vastly
inferior since the backend code in phpBB 3.0 doesn't do auth checking,
etc.

The code is a hybrid of the ideas from the KB article [1] for phpBB 3.0,
and the existing code in phpBB 3.1.

[1] https://www.phpbb.com/kb/article/adding-last-posts-subject-to-index/

Since phpBB 3.0 already reads the last post subject when generating
forum lists, this commit doesn't incur in additional query overhead.

Modifications for subsilver2, prosilver SE, wesnoth, and AcidTech (both
versions) included.

Cherry-pick note: actually a condensed patch composed of the following:

    1b1d668c0e95ff503da26a2b21adb603415e431d Use a plain ellipsis instead of an HTML entity when truncating post titles
    576fc7f5786a673857c7568a27598bca76f7d198 Truncate last post subject in forum lists using truncate_string()
    5ea66f6a234f87ebb82b3170237b070e7bea911e Revert "Wesnoth mod: unembolden last post titles in forum lists"
    7a9b407a759660a1453580b159046d3dae1d7d8c Wesnoth mod: unembolden last post titles in forum lists
    fd20afc7767a2b106ff8f3db11acb1f0615aaa2c Wesnoth mod: fix botched XML entity from previous commit
    f72736cde61f92fae346f74ef221aa3c62ef8123 Wesnoth mod: display last post subject in forum lists