Wesnoth mod: display last post subject in forum lists
This is a custom modification that's strongly derivative of the
functionality suggested for phpBB 3.1, although it's presumably vastly
inferior since the backend code in phpBB 3.0 doesn't do auth checking,
etc.

The code is a hybrid of the ideas from the KB article [1] for phpBB 3.0,
and the existing code in phpBB 3.1.

[1] https://www.phpbb.com/kb/article/adding-last-posts-subject-to-index/

Since phpBB 3.0 already reads the last post subject when generating
forum lists, this commit doesn't incur in additional query overhead.

Modifications for subsilver2, prosilver SE, wesnoth, and AcidTech (both
versions) included.